requirements.txt
openai==0.26.1
python-dotenv==0.21.1